+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Version 2.9.0.0
+
+- Added Discord Rich Presence support because why not
+- Added persistent raffles joined count
+  - Only raffles joined on or after this version of Scraps will be counted
+- Added status for joining poll
+- Fixed polls not working if they allow multiple choice
+- Chosen poll choice now shows the number as it appears rather than its programmatic number
+- Added pluralization to poll joining log lines
+- Updated app icon
+
 ## Version 2.8.0.0
 
 - Added option to vote in polls